+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # 1.10.4
 
+### Fixed
+
+- Fixed the Moon Phase Detector desync (#72)
+
 ### Changed
 
 - Removed message about Sodium incompatibility, because that was fixed in an Aaron update